bsp: atf: use ATF 2.0
This is a big update for ATF. It includes a lot of stability fixes as
well as new and exciting features like a proper core offlining and
making shutdown work more reliably.

The size is drastically bigger, though. More than 1.5 times. So in order
to make it fit to U-Boot SPL we need to make some adjustments to U-Boot
proper by either storing environment on a FAT partition (as in upstream)
or moving to 8Mb offset (because of the Mender partitioning).

I also propose to use a debug build since we don't really care about
size anymore. The output is actually of help so let's enable it.

PLAT is changed to sun50i_a64 because upstream ATF has a common target
for H5 and A64 because they are mostly the same. We'll need to make a
couple of adjustments to our neutis-n5.dtsi in order to make DRAM and
eMMC regulators do the right thing. And then we'll need to patch ATF to
support these changes.