bsp: atf: remove notimestamp
We used to have the problem that bl31.bin could not build,
because the 'dirty' u-boot flag would not allow the u-boot-upgrader
mechanism to run. Since then this problem has been fixed: we have
updated the u-boot-upgrader mechanism, u-boot naming has also been
changed to get rid of the dirty flag, and the ATF rebuild is now
unnecessary and only time consuming. Thus we should get rid of nostamp
to not rebuild ATF which leads to rebuild u-boot and rootfs
consequently.